+++ packages/firebase_database/android/src/main/java/io/flutter/plugins/firebase/database/FirebaseDatabasePlugin.java
@@ -8,8 +8,10 @@ import android.util.SparseArray;
 import com.google.firebase.database.ChildEventListener;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseException;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.Query;
 import com.google.firebase.database.ValueEventListener;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
@@ -31,7 +33,7 @@ public class FirebaseDatabasePlugin implements MethodCallHandler {
 
   // Handles are ints used as indexes into the sparse array of active observers
   private int nextHandle = 0;
-  private final SparseArray<EventObserver> observers = new SparseArray<EventObserver>();
+  private final SparseArray<EventObserver> observers = new SparseArray<>();
 
   public static void registerWith(PluginRegistry.Registrar registrar) {
     final MethodChannel channel =
@@ -43,12 +45,66 @@ public class FirebaseDatabasePlugin implements MethodCallHandler {
     this.channel = channel;
   }
 
-  private static DatabaseReference getReference(Map<String, ?> arguments) {
-    @SuppressWarnings("unchecked")
+  private DatabaseReference getReference(Map<String, Object> arguments) {
     String path = (String) arguments.get("path");
     DatabaseReference reference = FirebaseDatabase.getInstance().getReference();
     if (path != null) reference = reference.child(path);
     return reference;
+  }
+
+  private Query getQuery(Map<String, Object> arguments) {
+    Query query = getReference(arguments);
+    @SuppressWarnings("unchecked")
+    Map<String, Object> parameters = (Map<String, Object>) arguments.get("parameters");
+    Object orderBy = parameters.get("orderBy");
+    if ("child".equals(orderBy)) {
+      query = query.orderByChild((String) parameters.get("orderByChildKey"));
+    } else if ("key".equals(orderBy)) {
+      query = query.orderByKey();
+    } else if ("value".equals(orderBy)) {
+      query = query.orderByValue();
+    } else if ("priority".equals(orderBy)) {
+      query = query.orderByPriority();
+    }
+    if (parameters.containsKey("startAt")) {
+      Object startAt = parameters.get("startAt");
+      String startAtKey = (String) parameters.get("startAtKey");
+      if (startAt instanceof Boolean) {
+        query = query.startAt((Boolean) startAt, startAtKey);
+      } else if (startAt instanceof String) {
+        query = query.startAt((String) startAt, startAtKey);
+      } else {
+        query = query.startAt((Double) startAt, startAtKey);
+      }
+    }
+    if (parameters.containsKey("endAt")) {
+      Object endAt = parameters.get("endAt");
+      String endAtKey = (String) parameters.get("endAtKey");
+      if (endAt instanceof Boolean) {
+        query = query.endAt((Boolean) endAt, endAtKey);
+      } else if (endAt instanceof String) {
+        query = query.endAt((String) endAt, endAtKey);
+      } else {
+        query = query.endAt((Double) endAt, endAtKey);
+      }
+    }
+    if (arguments.containsKey("equalTo")) {
+      Object equalTo = arguments.get("equalTo");
+      if (equalTo instanceof Boolean) {
+        query = query.equalTo((Boolean) equalTo);
+      } else if (equalTo instanceof String) {
+        query = query.equalTo((String) equalTo);
+      } else {
+        query = query.equalTo((Double) equalTo);
+      }
+    }
+    if (arguments.containsKey("limitToFirst")) {
+      query = query.limitToFirst((int) arguments.get("limitToFirst"));
+    }
+    if (arguments.containsKey("limitToLast")) {
+      query = query.limitToLast((int) arguments.get("limitToLast"));
+    }
+    return query;
   }
 
   private class DefaultCompletionListener implements DatabaseReference.CompletionListener {
@@ -79,8 +135,8 @@ public class FirebaseDatabasePlugin implements MethodCallHandler {
 
     private void sendEvent(String eventType, DataSnapshot snapshot, String previousChildName) {
       if (eventType.equals(requestedEventType)) {
-        Map<String, Object> arguments = new HashMap<String, Object>();
-        Map<String, Object> snapshotMap = new HashMap<String, Object>();
+        Map<String, Object> arguments = new HashMap<>();
+        Map<String, Object> snapshotMap = new HashMap<>();
         snapshotMap.put("key", snapshot.getKey());
         snapshotMap.put("value", snapshot.getValue());
         arguments.put("handle", handle);
@@ -119,34 +175,54 @@ public class FirebaseDatabasePlugin implements MethodCallHandler {
     }
   }
 
-  @SuppressWarnings("unchecked")
   @Override
   public void onMethodCall(MethodCall call, final Result result) {
-    Map<String, Object> arguments = (Map<String, Object>) call.arguments;
+    Map<String, Object> arguments = call.arguments();
     switch (call.method) {
       case "FirebaseDatabase#goOnline":
-        FirebaseDatabase.getInstance().goOnline();
-        break;
+        {
+          FirebaseDatabase.getInstance().goOnline();
+          result.success(null);
+          break;
+        }
 
       case "FirebaseDatabase#goOffline":
-        FirebaseDatabase.getInstance().goOffline();
-        break;
+        {
+          FirebaseDatabase.getInstance().goOffline();
+          result.success(null);
+          break;
+        }
 
       case "FirebaseDatabase#purgeOutstandingWrites":
-        FirebaseDatabase.getInstance().purgeOutstandingWrites();
-        break;
+        {
+          FirebaseDatabase.getInstance().purgeOutstandingWrites();
+          result.success(null);
+          break;
+        }
 
       case "FirebaseDatabase#setPersistenceEnabled":
         {
-          boolean isEnabled = (boolean) arguments.get("enabled");
-          FirebaseDatabase.getInstance().setPersistenceEnabled(isEnabled);
+          Boolean isEnabled = (Boolean) arguments.get("enabled");
+          try {
+            FirebaseDatabase.getInstance().setPersistenceEnabled(isEnabled);
+            result.success(true);
+          } catch (DatabaseException e) {
+            // Database is already in use, e.g. after hot reload/restart.
+            result.success(false);
+          }
           break;
         }
 
       case "FirebaseDatabase#setPersistenceCacheSizeBytes":
         {
-          long cacheSize = (long) arguments.get("cacheSize");
-          FirebaseDatabase.getInstance().setPersistenceCacheSizeBytes(cacheSize);
+          Long cacheSize = (Long) arguments.get("cacheSize");
+          try {
+            FirebaseDatabase.getInstance().setPersistenceCacheSizeBytes(cacheSize);
+            result.success(true);
+          } catch (DatabaseException e) {
+            // Database is already in use, e.g. after hot reload/restart.
+            result.success(false);
+          }
           break;
         }
 
@@ -171,6 +247,13 @@ public class FirebaseDatabasePlugin implements MethodCallHandler {
           break;
         }
 
+      case "Query#keepSynced":
+        {
+          boolean value = (Boolean) arguments.get("value");
+          getQuery(arguments).keepSynced(value);
+          break;
+        }
+
       case "Query#observe":
         {
           String eventType = (String) arguments.get("eventType");
@@ -178,9 +261,9 @@ public class FirebaseDatabasePlugin implements MethodCallHandler {
           EventObserver observer = new EventObserver(eventType, handle);
           observers.put(handle, observer);
           if (eventType.equals(EVENT_TYPE_VALUE)) {
-            getReference(arguments).addValueEventListener(observer);
+            getQuery(arguments).addValueEventListener(observer);
           } else {
-            getReference(arguments).addChildEventListener(observer);
+            getQuery(arguments).addChildEventListener(observer);
           }
           result.success(handle);
           break;
@@ -188,14 +271,14 @@ public class FirebaseDatabasePlugin implements MethodCallHandler {
 
       case "Query#removeObserver":
         {
-          DatabaseReference reference = getReference(arguments);
+          Query query = getQuery(arguments);
           int handle = (Integer) arguments.get("handle");
           EventObserver observer = observers.get(handle);
           if (observer != null) {
             if (observer.requestedEventType.equals(EVENT_TYPE_VALUE)) {
-              reference.removeEventListener((ValueEventListener) observer);
+              query.removeEventListener((ValueEventListener) observer);
             } else {
-              reference.removeEventListener((ChildEventListener) observer);
+              query.removeEventListener((ChildEventListener) observer);
             }
             observers.delete(handle);
             result.success(null);